11: Added ConnectedCollectionWidget and DisconnectedCollectionWidget
